Add sitemap crawling CLI
