+++ linksInSitemap2pdf.py
@@ -1,0 +1,58 @@
+#!/usr/bin/env python3
+"""
+linksInSitemap2pdf.py
+---------------------
+Fetch a sitemap.xml file, crawl each URL listed, and compile the content into
+combined Markdown and PDF outputs.
+
+This CLI shares crawling and conversion logic with ``site2pdf_core``.
+"""
+
+import argparse
+import sys
+
+from site2pdf_core import (
+    normalize_url,
+    fetch,
+    extract_links_from_sitemap,
+    crawl_and_output,
+)
+
+
+def main():
+    ap = argparse.ArgumentParser(
+        description="Crawl all links in a sitemap.xml and build Markdown+PDF"
+    )
+    ap.add_argument(
+        "sitemap_url",
+        help="URL to sitemap.xml (e.g. https://example.com/sitemap.xml)",
+    )
+    ap.add_argument(
+        "--delay", type=float, default=0.35, help="Delay between page fetches (seconds)"
+    )
+    ap.add_argument(
+        "--limit",
+        type=int,
+        default=0,
+        help="Max linked pages to fetch (0 = no limit)",
+    )
+    args = ap.parse_args()
+
+    sitemap_url = normalize_url(args.sitemap_url)
+
+    print(f"[+] Fetching sitemap: {sitemap_url}")
+    status, content, _ = fetch(sitemap_url)
+    if status != 200 or not content:
+        print(f"[-] Failed to fetch sitemap ({status}).", file=sys.stderr)
+        sys.exit(2)
+
+    links = extract_links_from_sitemap(content)
+    if args.limit > 0:
+        links = links[:args.limit]
+    print(f"[+] Found {len(links)} link(s) in sitemap.")
+
+    crawl_and_output(sitemap_url, links, args.delay)
+
+
+if __name__ == "__main__":
+    main()
